+++ lib/t/t_ext2int2ext.m
@@ -15,7 +15,7 @@ if nargin < 1
     quiet = 0;
 end
 
-num_tests = 135;
+num_tests = 138;
 t_begin(num_tests, quiet);
 
 %% define named indices into data matrices
@@ -171,6 +171,7 @@ else
     t_is(mpc.branch, mpci.branch, 12, [t 'branch']);
     t_is(mpc.gen, mpci.gen, 12, [t 'gen']);
     t_is(mpc.gencost, mpci.gencost, 12, [t 'gencost']);
+    t_ok(isequaln(mpc.bus_name, mpci.bus_name), [t 'bus_name']);
     t_ok(isequaln(mpc.gentype, mpci.gentype), [t 'gentype']);
     t_ok(isequaln(mpc.genfuel, mpci.genfuel), [t 'genfuel']);
     t_is(mpc.A, mpci.A, 12, [t 'A']);
@@ -181,6 +182,7 @@ else
     t_is(mpc.branch, mpci.branch, 12, [t 'branch']);
     t_is(mpc.gen, mpci.gen, 12, [t 'gen']);
     t_is(mpc.gencost, mpci.gencost, 12, [t 'gencost']);
+    t_ok(isequaln(mpc.bus_name, mpci.bus_name), [t 'bus_name']);
     t_ok(isequaln(mpc.gentype, mpci.gentype), [t 'gentype']);
     t_ok(isequaln(mpc.genfuel, mpci.genfuel), [t 'genfuel']);
     t_is(mpc.A, mpci.A, 12, [t 'A']);
@@ -191,6 +193,7 @@ else
     t_is(mpc.branch, mpce.branch, 12, [t 'branch']);
     t_is(mpc.gen, mpce.gen, 12, [t 'gen']);
     t_is(mpc.gencost, mpce.gencost, 12, [t 'gencost']);
+    t_ok(isequaln(mpc.bus_name, mpce.bus_name), [t 'bus_name']);
     t_ok(isequaln(mpc.gentype, mpce.gentype), [t 'gentype']);
     t_ok(isequaln(mpc.genfuel, mpce.genfuel), [t 'genfuel']);
     t_is(mpc.A, mpce.A, 12, [t 'A']);